requirements.txt
asgiref==3.9.1
Django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.15.2
django-formtools==2.5.1
pillow==11.3.0
sqlparse==0.5.3
tree-cli==0.1.1
typing_extensions==4.15.0
gunicorn==20.1.0
whitenoise==6.7.0
django-storages==1.14.4
boto3==1.35.86
stripe==8.8.0
dj-database-url==2.2.0
reportlab==4.2.5
pytest==8.3.4
pytest-django==4.9.0
playwright==1.49.1
sentry-sdk==2.20.0
